Fix for linux
git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@13677 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.tests/examples/Benchmarks/UTS/sha1.c
@@ -37,7 +37,7 @@
 #include <stdio.h>
 
 #include "sha1.h"
-#include "endian.h"
+#include "sha_endian.h"
 
 #if defined(__cplusplus)
 extern "C"
@@ -116,7 +116,7 @@ char * rng_showstate(RNG_state *state, char *s){
 /* describe random number generator type into string */
 int rng_showtype(char *strBuf, int ind) {
   ind += sprintf(strBuf+ind, "SHA-1 (state size = %luB)",
-                 sizeof(struct state_t));
+                 (unsigned long)sizeof(struct state_t));
   return ind;
 }
